pyproject.toml
[project]
name = "emby-to-trakt"
version = "1.0.0"
description = "Sync watched history from Emby to Trakt"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "requests>=2.28",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "responses>=0.23",
    "ruff>=0.8.0",
]

[project.scripts]
emby-sync = "emby_to_trakt.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
